# internal/server/swagger/swagger.yaml
basePath: /
...
host: localhost:1324
info:
  contact: {}
  description: KNOV API \n http://localhost:1324
  title: Knov API
  version: "1.0"
paths:
...
  /api/config/datapath:
    get:
      produces:
      - text/html
      responses: {}
      summary: Get current data path as input field
      tags:
      - config
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: updates data path in .env file (requires restart)
      parameters:
      - description: data path
        in: formData
        name: dataPath
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: saved
          schema:
            type: string
      summary: Update data path
      tags:
      - config
...
  /api/config/favicon:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      description: Removes the custom favicon and reverts to the default
      produces:
      - text/html
      responses:
        "200":
          description: favicon removed
          schema:
            type: string
        "500":
          description: failed to remove
          schema:
            type: string
      summary: Delete custom favicon
      tags:
      - config
    post:
      consumes:
      - multipart/form-data
      description: Uploads a custom favicon (ico, png, or svg) stored in storage/favicon
      parameters:
      - description: Favicon file (.ico, .png, or .svg)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - text/html
      responses:
        "200":
          description: favicon uploaded
          schema:
            type: string
        "400":
          description: invalid file
          schema:
            type: string
        "500":
          description: upload failed
          schema:
            type: string
      summary: Upload custom favicon
      tags:
      - config
  /api/config/import:
    post:
      consumes:
      - multipart/form-data
      description: Uploads and applies user settings from a JSON file
      parameters:
      - description: Settings JSON file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - text/html
      responses: {}
      summary: Import user settings from JSON
      tags:
      - config
...
  /api/config/repository:
    get:
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get git remote URL
      tags:
      - config
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: updates git remote url in .env file
      parameters:
      - description: remote repository url
        in: formData
        name: repositoryURL
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: saved
          schema:
            type: string
      summary: Update git remote URL
      tags:
      - config
...
swagger: "2.0"